feat: Complete TaskMaster optimization project - All 15 tasks resolved with comprehensive documentation and production-ready enhancements
